requirements.txt
streamlit==1.32.0
openai==1.12.0
pyPDF2==3.0.1
faiss-cpu==1.7.4
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
python-dotenv==1.0.0
chromadb==0.4.22
bs4==0.0.1
requests==2.31.0
wikipedia==1.4.0
spacy==3.7.2
networkx==3.1
plotly==5.18.0
pandas==2.0.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
httpx==0.27.0
